Configure auto-deployment using Travis-CI for the master repo and branch
On merges into the master branch, the Docker image is pushed to Dockerhub
with the ":latest" tag. On releases, the Docker image is tagged with the
release tag and pushed to Dockerhub.

Exit with non-zero code if no DOCKER_USERNAME or DOCKER_PASSWORD present

Push images to ioos/ioos-us

Tag image appropriately

Clean up BASH script for Dockerhub push

Properly redirect stdout to stderr